Added cmake to the project, reorganized app/. Compiles
