+++ cmake/submodules.cmake
@@ -1,0 +1,1 @@
+# if ever adding submodules
